backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
pydantic==2.9.2
